flask-jwt/requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.0
mysql-connector-python==8.1.0
